myapp/requirements.txt
Django==3.2
psycopg2-binary==2.8.6
gunicorn==20.1.0
bcrypt==3.2.0
pillow==8.3.2
Faker==8.1.1
nplusone==1.0.0
pytest==6.2.4
factory_boy==3.2.0
python-dotenv==0.17.1
pytz==2021.1
pytest-factoryboy==2.5.1
pytest-django==4.5.2
django-filter==23.2
python-decouple==3.8
